Update AsyncTCP and Async WebServer libraries to current version since they contain vital fixes.  Switched from polling /status to pushing with WebSockets. Some ADC fixes. Reenabled OTA. Lots of refactoring and pesky bug fixes (like the io_acceleration ISR crashing)!

+++ src/battery.h
@@ -11,7 +11,7 @@ class Battery {
     Battery(IO_Analog& io_analog);
     float getChargerVoltage();
     float getBatteryVoltage();
-    float getBatteryStatus();
+    uint8_t getBatteryStatus();
     uint32_t getLastFullyChargeTime();
     uint32_t getLastChargeDuration();
     bool isCharging();
